requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
Django==3.1.3
django-environ==0.4.5
djangorestframework==3.12.2
idna==2.10
psycopg2-binary==2.8.6
pycparser==2.20
python-telegram-bot==13.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
